AIAim class for handling AI looking at targets
I created a new MonoBehaviour class, AIAim, to control how the AI aims and looks at targets.
This code revolves around a Quaternion value named lookRotation, that determines the current direction the AI is meant to be looking in. This Quaternion is held in code instead of having a child transform representing the look direction like with the player controller, so the rotation doesn’t awkwardly shift around when the AI’s NavMeshAgent turns corners.
Functions are also present for rotating the AI’s look direction through lookRotationb, and checking if it’s close enough to a target to be looking at it. Several Vector3 properties are present to easily obtain direction values from the Quaternion.
All of this code already existed in the AI class, but I moved it over for improved modularity, as well as making some changes.
For example, the aim sway code I included in WeaponHandler has been moved to this class as well, and turned into a static function. This will make it easier to use this aim sway code for enemies as well as the player. Instead of simply returning a Vector3 direction, it now returns a Quaternion representing an angle deviation, which is then multiplied onto the player or enemy’s look direction and then multiplied by a Vector3 to determine the current aim direction.
I also made an AimValues struct that holds data for looking around, such as aim speed, a detection mask, sway variables and a diameter required for unobstructed line of sight. The AIAim class holds a default reference to one, but it can be swapped out based on the current desired AI behaviour.
Finally, I also added a second look checking bool, that uses the distance between the AI and the target to determine how far away in distance its aim is. This will be better for looking at objects with a variable size or distance, since an object that’s further away will require a stricter angle calculation.