EngageTargetAtDistance behaviour
This new AIAction-derived class contains code to have an enemy pursue the player while also maintaining a safe distance. When the state is entered, the enemy will run a function named FindIdealLocation(). This will search for a certain number of locations around the player, within a minimum and maximum range, and search for the closest position that the AI can reach where they can establish line of sight with the player.
During the update loop, the enemy will make its way towards the ideal location, continuously checking that the target position is still viable (i.e. that it is not too close or too far away from the target and that line of sight can still be established). If not, FindIdealLocation() will run again to find a new location to move to.

Since I have designed the player and enemies to have their transforms at the bottom so they can easily be placed on surfaces, I had to devise some extra code to ensure the line of sight checks aim for the right location.
I created a function named NewPositionLookOrigin, which takes the destination value and combines it with the relative position difference between the enemy’s transform and their ‘head’, to produce a position representing where the enemy’s eyes would be if they were standing at that position.
I created a series of functions named CombinedBounds, that return a Bounds struct encapsulating an array of Colliders, in this case the player hitboxes. The line of sight then aims towards this struct’s centre position.
My line of sight check function runs a SphereCastAll (instead of a RaycastAll, to ensure the sightline is actually wide enough for the enemy to look through) between an origin position and an array of target colliders, and sorts the results by distance. It then checks through each returned collider to see if it’s part of the target (returning true), part of a list of exception colliders such as the enemy’s own body (ignoring), or neither (returning false since it’s actually something obstructing the view of the player). In order to easily obtain the player’s colliders as an array, excluding their hitbox data, I added a HitboxColliders property that generates and stores an array of all colliders used as damage hitboxes for a character, and returns it on each subsequent call.

Overall, I am very pleased with how this code has turned out. The test enemy I created does a fantastic job of pursuing the player and being able to ‘aim’ at them, without getting too close, as a real combatant would if armed with a ranged weapon. Being able to easily adjust the minimum and maximum distance can allow me to tailor the behaviours for many different enemy types, and adjusting the number of checks lets me easily control how precise the pathfinding is.