AimAtTarget class for enemy targeting
I made a new behaviour class that inherits from AIAction, where the enemy will aim at a target.
This behaviour makes heavy use of the AIAim class. If the enemy is capable of establishing line of sight (checked using AIAction.LineOfSightCheck()), they will continuously rotate their aim towards the player. A bool named ‘targetAcquired’ is updated based on if the enemy’s aim is successfully on target. Two separate threshold values are included (aimLockThreshold and aimBreakThreshold), so the AI aiming can have a bit of stickiness after they lock onto a target.
Three UnityEvents are present – onTargetAcquired, whileTargetAcquired, and onTargetLost – which activate when an accurate lock is first obtained, every frame while the lock is valid, and when the lock is broken.
While my test enemy does not presently perform any actions on targeting a player, this aiming code works well and I am very proud of it.