Fixed AIAim distance check bool not calculating positions correctly
When calculating the position out in the direction that the enemy is aiming, at the same distance as the target, I forgot to add the look origin position to the direction value. This was causing the positions to be calculated incorrectly, as if the enemy's look origin was (0,0,0). This has now been fixed.