Tweaked AIAim functions to rotate enemy head and body
I added some extra functionality in AIAim, so when an enemy adjusts their aim their head and body will visibly turn to match.
Rotating the head is simple enough, as I can directly set its rotation to the lookRotation Quaternion. However, since the head rotates relative to the body, and the body’s rotation would normally be perpendicular to the surface it’s standing on, the body obtains a version of the current look direction as a Vector3, ‘flattened’ using Vector3.ProjectOnPlane so it is perpendicular to the body’s up axis, while otherwise facing the same direction. This allows the enemy’s head and body to rotate just like the player does.
Additionally, I tweaked the ‘ReturnToNeutralLookPosition’ function to look in the direction of the NavMeshAgent’s current velocity, rather than simply straight forward. This will have the effect of the AI properly looking in the direction they’re moving, if they are walking up or down a slope or staircase.