Friendly fire avoidance for bullets
My PointDamage class for dealing bullet damage had a bool named ‘notAlly’, that detects if the target is safe to shoot. By default, this was always set to true, with the expectation that I would add in actual friendly fire code later on. Now, I have. The class will quickly check for a Character class in the targeted collider or one of its parents, and check the affiliation against the entity that launched the initial projectile. If the attacker and target are allies, the target will not be damaged.