Tweaked AIAim look shifting functions to require less parameters
I changed the RotateLookTowards() and ReturnToNeutralLookPosition() functions to not require a speed parameter, as that value is already specified in AIAim’s ‘stats’ struct. This will make writing said code easier, with less opportunities for typing incorrect code and causing bugs.
I also renamed ReturnToNeutralLookPosition() to LookInNeutralDirection(), as the former implies it’s a function you run once, whereas the latter implies its true usage as something that runs continuously.