AIAttack class for enemy attack sequences (untested)
I created a new MonoBehaviour class, AIAttack, that includes code for performing sequences of enemy attacks, including telegraphing and cooldowns.
The main element of the class is an IEnumerator that queues up a telegraph, a number of attacks and a cooldown. This is then managed using a bool property that starts and ends the attack, UnityEvents that run while telegraphing, attacking and cooling down, and public functions to easily activate other forms of attack (such as a ShootGun() class that will fire shots from a GunGeneralStats class somewhere on the enemy).