Implementing AIAttack into AimAtTarget
I had to overhaul much of my AimAtTarget class in order to make it work with AIAttack. Most if it was rather dry technical stuff, but the major changes are listed below:
Added a better check for determining if the AI is successfully aiming at a target. Since the aiming code already uses bounds to determine the centre of mass to aim at, the check will use the dimensions of the enemy’s volume to determine an acceptable aim threshold. If the player is already aiming, an aim break threshold value is added on to give the aim a degree of ‘stickiness’.
Tweaked code for determining the point in space to aim at, and whether the AI is successfully aiming at a target, to work off two virtual functions that can be altered in inherited classes. This is so if I need an enemy that attacks in a way that’s very different from simply requiring line of sight, I can make it easy for the enemy to aim towards the right target.
Removed UnityEvents that occurred when acquiring aim, while aim was acquired and when aim was broken, since I was not using them for anything.