Combatant class – adding code for when outside of combat
I started adding some code for when an AI combatant is not presently attacking anything.
Added a new SubStateMachine named ‘outOfCombat’ for when the AI is not presently attacking anything. At this point, the AI will just idle.
Added a new MultiState named ‘inCombat’, so I can assign movement and attack functions in a single state.
Tweaked the assignment of states in BasicGunEnemy to be assigned to inCombat, rather than the base state machine.
Added state transitions to switch between the ‘inCombat’ and ‘outOfCombat’ states based on the TargetAcquired() Func<bool>.
Tweaked the TargetAcquired() transition criteria to have a bool parameter named ‘condition’, allowing the transition to check whether the criteria is true or false. The returned bool exists to check if it’s appropriate to transition, but in some cases it might be appropriate to transition if the value returns false rather than true.
I also altered the AI class to only require one StateMachine, as the new MultiState and SubStateMachine classes act as a much better way to run multiple states simultaneously.