Tweaked gun firing code so bullets don’t get stuck in the shooter’s own colliders
I had to fix this bug when I noticed that my placeholder enemy’s bullets would hit its own colliders. The raycasts I used to determine where to rotate the spawned projectile towards have been replaced with using RaycastAll to determine the target location as well as the enemy’s colliders, and excluding all colliders that are part of the enemy’s body. This code also works for the player.