AI will now automatically look in their neutral direction if they aren’t actively aiming towards something.
I had to fix this because I noticed enemies could get stuck looking in a particular direction if they didn’t have an acquired target. This code is disabled when the AI switches to a state that involves its own aiming code, and is enabled when the AI switches back out of said state.
I also tweaked the code for determining the direction the AI looks in. By default it uses the NavMeshAgent velocity to ensure the AI looks in the direction they’re moving, but if they aren’t moving this value is zero so there’s no valid direction value. If so, it rotates to look straight forward in the current direction its body is facing.