Fixed bug where enemy attack cooldown didn’t initiate properly
I noticed that my enemy was attacking unusually fast. I realised that this was because their attack was triggering again before the cooldown was finished. The code I used to handle showing the current state of the attack was very messy, and it made sense that it would cause bugs, so I overhauled it. The new system uses an enum to determine the current status of the attack, with four states – Ready, Telegraphing, Attacking and CoolingDown. By using this enum to determine the current state of the attack, it’s much easier to have the other mechanics behave accordingly with no glitches.