Work on CombatArea class to easily control enemy aggression
This new class is attached to a trigger in the game level, that encompasses an area in which a fight takes place. When the player enters the trigger, all the enemies within the area become aware of them, and they will become peaceful again when the player leaves. This acts as a simplistic but quick and effective way to initiate a combat encounter.
It needs work (such as being able to perform custom actions upon entering, leaving or completing a fight), and it won’t be ideal for all scenarios, but I am very pleased with how it’s turned out so far.