Bug fix: weapons and ADS continue to happen even if the player dies
This effect would look very strange if the player died while shooting and/or aiming down sights. I have now fixed this so those abilities are automatically cancelled when the player dies. This is activated by a public CancelInputs() function, which I can trigger from anywhere, so it will also be useful for other effects like being stunned or forced into a cutscene.