Door object
I created a simple Door class, that opens and shuts when a bool is switched, The opening and shutting is handled using UnityEvents, so I can assign whatever actions I want to something that might not resemble a regular door but would function as such, e.g. a force field. This system is not perfect and will probably be refined or overhauled in future, but it works as a temporary feature to allow gating of level areas.