Bug fix: Tweaked torso drag animation on player controller, so the direction stays consistent regardless of where the player is looking.
I noticed that when the player is looking upwards or downwards while moving, the movement of the weapon model will still move as if they are facing straight forward. E.g. the model moving inwards while walking forwards or outwards while walking backwards. I fixed this by changing the transform used as reference to convert the direction values, from the player's main transform to their head transform, since the torso object that moves around is a child of the head.