Fixed bug where player camera would not match forward facing direction, while looking directly up or down
I attributed this to a bug in the GunADS code, that uses Transform.LookAt to rotate the upper body transform to one offset from the aim axis. However, it uses the up axis of the player’s movement transform rather than their aim axis, so if the player was looking directly up or down while not aiming down the weapon sights, the code would get confused due to the desired forward and up axes being the same.