Directional HUD indicator
I created the DirectionalHUDIndicator class, for a swivelling HUD indicator that shows the direction of an off-screen position. While active, the class will check the position of a specified transform, character hitbox bounds, or Vector3 position, and calculate the direction to that target. After checking that the angle is greater than a desired threshold (to ensure the arrow doesn’t appear for stuff too close to the reticle that the player can obviously see), the direction is ‘flattened’ to the player’s aim direction (and therefore the axis of the player’s HUD) using Vector3.ProjectOnPlane(), then converted to local space and used as a direction value to position and rotate the image on the player’s HUD canvas.
This system also includes a reference to a TimedCosmeticEffect, in order to play simple animations and for the visual aid to destroy itself after the animation has finished.
I am very pleased with how this system is turning out.