Changes to TimedCosmeticEffect - adjusting playback in editor and determining if an animation is complete
I tweaked this class to not reset the timer to zero upon finishing the animation. I added a bool property named 'completed' to tell if the animation is finished or not. I also made the timer into a public value alterable with a slider, so the player can adjust how the animation looks in the Inspector.