Directional damage indicators
Using the new DirectionalHUDIndicator class, I devised a system to allow the player to easily tell if they are being attacked from an offscreen direction.
The code involves taking the health and damage display code from the HeadsUpDisplay class and transferring into a new MonoBehaviour class, for improved modularity and cleaner code. I then added code to track a list of DirectionalHUDIndicators attached to the player, as well as a prefab one for a damage indicator arrow. Whenever the player takes damage, a new indicator arrow will spawn to show the direction of the attack source. If an indicator is already present for the most recent attacker (e.g. if they are attacking multiple times in quick succession), that indicator will simply be reset instead of making a new one, to save memory by not rapidly creating and destroying multiple arrows.
This system might need some minor improvements, but is overall a massive step forward for informing players about attacks they might not immediately see or hear.