Tweaked DirectionalHUDIndicator so it can be set to appear no matter what the viewing angle is
I made this change firstly by reducing the inspector range for the minimumAngleToShow value from 10 to 0. I also changed the outsideThreshold check from a greater than equation (>) to a greater than or equal to (>=) one, so even if the angle is as low as it can be, it can be counted as true.