Started work on improved AI attack system
My current AI attack code has a big problem in that it’s split between two classes in a very messy and awkward way. There’s an AimAtTarget state that shifts the AI’s aim towards the target and handles aiming code, and an AIAttack class that contains the stats and sequence code. Since this is too cluttered for my liking, I am working on combining them together into a single class named AIAttackBehaviour.
This abstract class is a state machine state containing some base code like the attack sequence and timing code, as well as a virtual bool function to determine if it is viable for the enemy to initiate the attack. The code currently checks the cooldown status and that the sequence is not already in progress. It also contains an abstract function for the AI to run additional code for performing an attack (e.g. for a ranged attack, it will shift its aim towards the target).
There’s also an inherited class named AIRangedAttack, which includes code such as adjusting the enemy’s aim speed while telegraphing, additional criteria for determining if the attack can be performed (such as the enemy aim being in the correct range and angle threshold, and line of sight being established).
These classes are not tested and may need more tweaks, but I am more pleased with these than with my current attack system.

I also slightly tweaked how the look check bools in AIAim work. Previously they involved specifying a separate Vector3, which was used to determine if the AI was merely looking at a target or specifically aiming at it. I changed this so that difference is instead determined by using a bool.