Movement controller now tweaks direction when walking down a slope, to reduce ‘skipping’
Previously, when moving, if the player moved too fast down a slope that is too steep, they will run right off, become airborne, and fall. This produced an awkward ‘skipping’ effect. I fixed this with code that obtains the normal direction of the ground the player is standing on, and shifts the player’s desired movement direction using Vector3.ProjectOnPlane, so it aligns with the surface normal. This ensures that the player moves relative to the angle of the surface, and will run down a hill they’re standing on instead of falling down it.