Work on AI attack avoidance
Since my enemies have very little in the way of self-preservation, I want to add code so they can detect if they are in the path of an attack, and avoid it.

I started work on an AttackMessage class, to be transmitted via the EventHandler whenever a player or AI character performs an attack. I also worked on an AvoidAttack state that the enemy uses to receive an incoming AttackMessage and find a safe location. I also made some additional tweaks such as changing the complex field of view detection check to also return angle and range checks, since those are performed automatically as part of the larger function and saving those values will mean the AI won't need to waste processing power calculating them again in the same frame.

This code is nowhere near complete, but I am pleased with how it's going.