Finished implementing AIAttackBehaviour
I converted the class into a MonoBehaviour so I could easily attach it to separate objects for the purpose of editing stats.
Created a state named ExecuteAttack, for the purpose of running the AIAttackBehaviour. This might seem strange, since a big problem with my previous system was having the code linked between two separate classes, but ExecuteAttack contains only the bare minimum of code while all the important functionality is still in AIAttackBehaviour.
Added virtual Enter() and Exit() functions to the AIAttackBehaviour class. For example, in AIRangedAttack, the code is used to disable and enable the AI’s default look direction, for proper manual aiming.
Overhauled my current enemy to use the new system.
These changes may seem like unnecessary fuss, but they have made my code much cleaner and easier to work with, which will help with making new attacks and adding them to behaviours down the line.