Tweaked 'Player' and 'AI' to both inherit from Character and share several abstract variables
In order to make my AI code work consistently when attacking the player or other NPCs, there are several factors that they share, which there was no easy way to account for. These factors included their ‘head’ or the place they look and aim from, and the layer masks used for looking and attacking. I changed ‘Character’ to an abstract class, and changed ‘AI’ to inherit from it. I then added several abstract properties, ‘LookOrigin’, ‘lookMask’ and ‘attackMask’, which ‘Player’ and ‘AI’ override in their own ways to account for their different systems.
Having an easy way to obtain, for any ‘character’, these properties, will be very useful for the different kinds of attack avoidance code I am working on, so my enemies can have a degree of self-preservation.