Reworked CombatArea to not rely on its own collider
Having a giant area-encompassing collider attached to the combat area was causing problems with AI grid point code, and with enemies being children of said area. In the bunker level, the encounter is now activated by the player using PlayerEventTrigger, passing through a trigger when they walk into the only entrance.