Reworked AIGridPoints to differentiate between open spaces and those next to cover
I created a new struct, GridPoint, which contains not only a Vector3 for position, but a bool to determine if it is cover or not, i.e. if it is close to a wall. This is calculated during generation with a usage of Physics.OverlapBox.
I’m making this change because I want enemies to be able to stay close to cover even when they’re exposing themselves to attack the player. Currently the enemies have no qualms about running right out into the open to attack the player.