EngageTargetAtDistance allows enemies to stay close to cover
This function now has a bool named ‘stayCloseToCover’, which adds some extra checks when searching for a new position.
If a position is confirmed to meet all the other requirements, the class will also check for nearby cover points. If there is one in a sufficient radius, angled so that it will actually protect from the player’s attacks, a bool will be marked prioritising it over positions that aren’t close to cover.
This code has helped ensure that my enemies look smarter and more cautious, by staying close to defensive positions instead of rushing out into open spaces like idiots.