Func<bool> for if AI cannot currently engage target
If an AI running EngageTargetAtDistance finds its position is no longer ideal for attacking its target, it will find a new position. But sometimes it can't do so. I changed this code slightly so results for if the current target is compromised, and if a new position can be found, are stored as bools. A Func<bool> has been implemented that checks these values to confirm if the AI cannot find an appropriate position to engage the target from. This will be so that the AI can take alternate actions, e.g. throwing a grenade to flush out an overly defensive player.