Updated weapon handler initialising function to appropriately assign parent and transform position
The code will now properly parent each new weapon to the assigned ‘socket’ transform, and reset the positions and rotations. This means I can easily drag a weapon into any part of the player transform and it will be automatically equipped and have its appropriate positions assigned.
I also condensed this code and the code for adding icons to the weapon selector to use a single for loop, rather than the two used previously.