Stamina and stun mechanics
The new Stamina class is a MonoBehaviour that stores a secondary ‘stamina’ meter as well as your health meter.
When an attack is performed, it has a ‘stun’ stat. When a character takes damage, if they have this class attached, their stamina will be worn down based on the ‘stun’ value, giving them a small stumble if the attack is over a certain threshold.
When depleted, an enemy will be forced into an extended stagger animation, disrupting their current behaviour and allowing the player to knock them about somewhat. When this happens, their stamina meter is refilled and they are temporarily rendered immune to stunning, to allow them a chance to bounce back and making it harder for the player to cheese fights by stunlocking enemies.