HumanoidAnimator class
This MonoBehaviour script exists to control animations on a humanoid rigged character model, by taking data from a Character class. It presently accomplishes two main functions.
Firstly, it updates the movement states based on their movement direction relative to the direction they’re facing.
Secondly, it applies offset rotation so their upper body points in the direction they’re aiming without interfering with the animations. This is accomplished by calculating the relative rotation between the character’s transform direction and their aim direction, then multiplying it onto the spine bones so that their existing positions due to animation are not overridden. A for loop and Quaternion.Slerp() is used to apply weight to the array of spine bones, so the rotation smoothly transitions over the whole spine.