Polished handgun firing animation to be smoother
Previously, the handgun's shooting animation was the exact same length as the delay for the next shot. Due to the incredibly quick firing speed, this resulted in an incredibly short, choppy animation. I changed it to be longer and smoother, plus I added a new transition to allow each shot to flow into the next despite overlapping between shots. This has made it look a lot more aesthetically pleasing.