Tweaked 'Combatant' class to be a separate behaviour from AI, rather than inheriting from it.
I made this change because even though it contains some data for remembering and pursuing a target, its main purpose is simply to set up logic behaviours, since I haven't created a GUI for it yet. By making it a separate class, I can easily remove it once I no longer have a use for it.