Implemented ragdoll activation after death
To enable a character's ragdoll physics, my original idea was to simply enable the script through a UnityEvent. But this caused several glitches, due to the rigidbodies still being parented to the base object.

I fixed this by adding some extra code to the Character class that references its ragdoll, and unparents it upon activation when the character dies. This isn't a perfect system and doesn't account for other times a character may need to be in a ragdoll state, but it works very well for death animations.