Replaced ‘EventHandler’ class with ‘Notification’ class
The old EventHandler class was problematic in that for each type of ‘event’ I wanted to transmit, I had to essentially re-type the same code again with the variables changed.
The new ‘Notification’ class uses a generic T variable, which eliminates that issue. For example, the System.Action that functions subscribe to is `public static System.Action<T> Receivers;`. This way, a function that wants to receive a certain value can subscribe using something like `Notification<SomeType>.Receivers += FunctionUsingSomeType;`, and other functions that want to transmit certain values can run a function like `Notification<SomeType>.Transmit(variableOfSomeType);`. This generics-based system means I do not have to rewrite any of this code for any new message I want to transmit, and also has the unintended but interesting side effect of being able to transmit any kind of variable this way.

This new system is much cleaner and more modular than the old EventHandler system, and I’m very pleased with how surprisingly well it turned out. I’m also pretty surprised with, for all the effort I put in, how deceptively simple the final solution turned out to be.