Overhauled AttackMessage class for cleaner and clearer code
I performed some work on an AttackMessage class, to clean up the code and make it better-written and more intuitive compared to last time.
I also created a new class, AOEAttackMessage, that contains its own check for an attack that occurs in a specified blast radius such as a grenade.