Reordered Damage() logic in Health class
I tweaked the DamageMessage transmission to only occur if the enemy is actually damaged, as previously it would also happen if they were healed. I also made sure it would still occur even if the enemy dies, because previously it wouldn’t.
I also tweaked the IsAlive property to be more compact and easier to read with less unnecessary whitespace.