MoveToDestination abstract class
I created a new class inheriting from AIMovement, designed for actions involving moving to a specific position based on certain criteria.

This class is designed to be easily slotted into a PriorityActionController, using a handy CanAndShouldMove() function that can simultaneously determine if there’s a reason to move, if there’s a valid place to move to, and if a previously valid destination is now invalid.

I have created this function because I realised that my EngageTargetAtDistance and AvoidAttack functions use a lot of similar logic, save for the specific criteria by which locations are found, so to this end I have decided to make this new class and have them both inherit from it, to keep the code consistent.