Added checks to AvoidAttack to prioritise positions close to cover, and which result in the least damage being taken.
It might be inevitable that an AI takes some damage while attempting to avoid an attack, so they will still try to take as little damage as possible.
Additionally, an AI character might not be brave enough to simply dodge an incoming attack, and may choose to be more cautious and seek valid cover first.