IPriorityAction interface, to easily pair actions and prerequisites for a PriorityActionController
Many Action classes in this project are designed to be used in a PriorityActionController, with a specific Func<bool> as the prerequisites. By creating the IPriorityAction interface, I can easily pair off actions and prerequisites, so they can be added in one easy package, using overloads of the PriorityActionController's AddAction and InsertAction functions.

In the future, I may be able to add similar interfaces for Actions designed to work in other types of controllers, e.g. a GOAP system.