‘cautionMultiplier’ threshold for enemy attack avoidance
This float determines how the enemy perceives the range of an attack. If this number is equal to one, the enemy will know the exact dimensions of the attack zone. If greater, they will play on the safe side and not get too close to an attack even if they might potentially be safe. If lesser, they will underestimate the zone of an attack and be riskier. When the PositionAtRisk bool is run, the dimension variables making up the zone of the attack are multiplied by the value, making the enemy’s perception of the danger zone larger or smaller.
I implemented this change because in my previous FPS project, which used similar avoidance code but lacked this variable, enemies would only dodge if they were precisely in the path of an attack. This meant that they would act too casually around attacks with very small but hard to judge danger zones. For example, it was not uncommon for enemies to not respond to having a bullet shot right next to them, because they technically weren’t inside the attack zone. As my enemies would realistically be much more cautious about being that close to gunfire, I wanted to solve this problem, and I feel this change will be a very elegant solution.