Enemies now register new attacks to avoid
Since AvoidAttack is not a MonoBehaviour, and is created and assigned as a variable to an AI’s ActionExecutor, I was wondering how to have it react to new attacks when the state itself is not active, in a way that is also clean and modular.
My method was to make a void function named CheckIncomingAttack, which is assigned to the Notification class’ receivers during AvoidAttack’s Setup() function. CheckIncomingAttack uses an AttackMessage as a parameter, and checks three things. First, it ensures the host controller is enabled and therefore that the AI is active and the enemy is alive and perceiving its surroundings. Second, it checks if the ‘currentAttackToAvoid’ variable is null, meaning that the enemy is not distracted avoiding a different attack. Finally, it checks if the enemy is actually in danger from the attack. If so, the new AttackMessage is assigned to ‘currentAttackToAvoid’, therefore enabling ReasonToMove() and triggering the enemy to avoid the attack.