Added attack avoidance stats to placeholder enemy, for testing purposes
The results are promising but not perfect. The enemy will avoid attacks with a degree of caution and realism, as opposed to ignoring everything that won't technically hit them. However, things get strange after the enemy first avoids an attack, and it's inconsistent whether they will avoid a second attack or not. More testing and troubleshooting are required.