GunReticle class for expanding reticle
This class uses a reference distance to visually indicate the spread for each kind of weapon. It also contains the code for determining when to show/hide the reticle, which means I can now remove it from HeadsUpDisplay, as part of splitting up the code in said class into separate and more appropriate classes.