Fixed weapon icons for radial menu
I added the necessary code for rotating the icons to be the right way up, if the initial rotation means they're upside down.
I also improved the visual design of the wheel, since superimposing the weapon graphics behind the other elements looked really bad. Since I realised the GUI elements were extremely large for a game that would not need to be played on a small screen, I shrunk most of the elements to fit a second, inner wheel that the weapon icons are now showed on. I changed the size of the dividers to separate this inner wheel as well, neatly showing what gun has what modes.