RecoilController bugfixes – cancel now doesn't cause jerks due to aim clamping
If the recoil caused the player's aim to go past the limit, and the player shifted their aim to cancel the recovery, the extra degrees would be lost via clamping and the player's vertical aim would jerk suddenly, looking very ugly.
Now, if the player shifts their aim to cancel the recovery effect but the ‘jerk’ would happen, it only cancels the horizontal recoil, retaining the vertical recoil until the player’s aim would no longer exceed the clamp value.