Look controller refactoring - improved gamepad aim acceleration and simplified control options
I made a lot of changes to the aim acceleration, including altering the sensitivity, the curve by which the value increases, and ensuring the acceleration only applies if the strength of the player’s input goes past a certain threshold (so it doesn’t apply if the player is only making slight adjustments). I also removed the ‘aimAcceleration’ and ‘timeToMaxAimAcceleration’ values, as these values are now stored in the ‘aimAccelerationCurve’ animation curve.
I also changed the way the sensitivity options are changed. Instead of directly changing the sensitivity values, there are now private ‘base’ values multiplied by ‘range’ integers, which are the values the player actually alters in the options menu. This ensures I can make whatever tweaks I want to the sensitivity in the player prefab, without having to alter how the options menu will change said values, because the ‘range’ values already represent the value that is displayed for the user to alter.