Shotgun balancing - increased damage but reduced accuracy
The shotgun felt pretty weak at close range so I increased the damage, now if all the pellets connect it can kill a standard target in a single shot. This made it ridiculously effective at long range, so I greatly increased the spread to make it harder for all the pellets to hit at longer distances.