Refactored gun fire controls to be more self-contained
The code jumped back and forth between RangedAttack and GunFireController, making it difficult to determine the flow of logic.
Additionally, I wanted to take turn SingleShot() into an IEnumerator and incorporate the fire delay, so I can make the RangedAttack code work more effectively with AI enemies that use guns.